% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.R
\name{confusion.matrix}
\alias{confusion.matrix}
\title{confusion.matrix}
\usage{
confusion.matrix(newdata, prediction)
}
\arguments{
\item{newdata}{matrix or data frame of test data.}

\item{prediction}{a prmdt prediction object.}
}
\value{
A matrix with predicted and actual values.
}
\description{
create the confusion matrix.
}
\examples{

data("iris")

n <- seq_len(nrow(iris))
.sample <- sample(n, length(n) * 0.75)
data.train <- iris[.sample,]
data.test <- iris[-.sample,]

modelo.knn <- train.knn(Species~., data.train)
modelo.knn
prob <- predict(modelo.knn, data.test, type = "prob")
prob
prediccion <- predict(modelo.knn, data.test, type = "class")
prediccion
confusion.matrix(data.test, prediccion)

}
