% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{contr.dummy}
\alias{contr.dummy}
\title{contr.dummy}
\usage{
contr.dummy(n, contrasts = TRUE)
}
\arguments{
\item{n}{A vector containing levels of a factor, or the number of levels.}

\item{contrasts}{A logical value indicating whether contrasts should be computed.}
}
\value{
A matrix with n rows and n-1 columns for contr.ordinal, a matrix with n rows and n columns for contr.dummy and a vector of length n for contr.metric.
}
\description{
Returns a matrix of contrasts for the \code{\link[kknn]{train.kknn}}.
}
