% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{dummy.data.frame}
\alias{dummy.data.frame}
\title{dummy.data.frame}
\usage{
dummy.data.frame(
  data,
  names = NULL,
  omit.constants = TRUE,
  dummy.classes = getOption("dummy.classes"),
  all = TRUE,
  ...
)
}
\description{
function from dummies package
}
\keyword{internal}
