% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{plot.prmdt}
\alias{plot.prmdt}
\title{Plotting prmdt models}
\usage{
\method{plot}{prmdt}(x, ...)
}
\arguments{
\item{x}{A prmdt models}

\item{...}{optional arguments to print o format method}
}
\value{
a plot of a model.
}
\description{
Plotting prmdt models
}
