% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ada.prmdt}
\alias{predict.ada.prmdt}
\title{predict.ada.prmdt}
\usage{
\method{predict}{ada.prmdt}(object, newdata, type = "class", n.iter = NULL, ...)
}
\description{
predict.ada.prmdt
}
\references{
Mark Culp, Kjell Johnson and George Michailidis (2016). ada: The R Package Ada for Stochastic Boosting.
R package version 2.0-5. https://CRAN.R-project.org/package=ada
}
\keyword{internal}
