% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.adabag.prmdt}
\alias{predict.adabag.prmdt}
\title{predict.adabag.prmdt}
\usage{
\method{predict}{adabag.prmdt}(object, newdata, type = "class", ...)
}
\description{
predict.adabag.prmdt
}
\keyword{internal}
