% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gbm.prmdt}
\alias{predict.gbm.prmdt}
\title{predict.gbm.prmdt}
\usage{
\method{predict}{gbm.prmdt}(
  object,
  newdata,
  type = "class",
  n.trees = NULL,
  single.tree = FALSE,
  ...
)
}
\arguments{
\item{object}{a \code{\link[gbm]{gbm}} model object for which prediction is desired.}

\item{newdata}{an optional data frame in which to look for variables with which to predict.}

\item{type}{type of prediction 'prob' or 'class' (default).}

\item{n.trees}{Number of trees used in the prediction. n.trees may be a vector in which case predictions are returned for each iteration specified}

\item{single.tree}{If single.tree=TRUE then predict.gbm returns only the predictions from tree(s) n.trees.}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
a vector or matrix of predictions gbm model.
}
\description{
Return prediction for a \code{\link[gbm]{gbm}} model.
}
