% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.glmnet.prmdt}
\alias{predict.glmnet.prmdt}
\title{predict.glmnet.prmdt}
\usage{
\method{predict}{glmnet.prmdt}(object, newdata, type = "class", s = NULL, ...)
}
\arguments{
\item{object}{a \code{\link[glmnet]{glmnet}} model object for which prediction is desired.}

\item{newdata}{an optional data frame in which to look for variables with which to predict.}

\item{type}{type of prediction 'prob' or 'class' (default).}

\item{s}{a \code{\link[glmnet]{cv.glmnet}} object (optional).}

\item{...}{additional arguments affecting the predictions produced.}
}
\description{
Return prediction for a \code{\link[glmnet]{glmnet}} model.
}
