% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.knn.prmdt}
\alias{predict.knn.prmdt}
\title{predict.knn.prmdt}
\usage{
\method{predict}{knn.prmdt}(object, newdata, type = "class", ...)
}
\description{
predict.knn.prmdt
}
\references{
Klaus Schliep and Klaus Hechenbichler (2016). kknn: Weighted k-Nearest Neighbors.
R package version 1.3.1. https://CRAN.R-project.org/package=kknn
}
\keyword{internal}
