% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.neuralnet.prmdt}
\alias{predict.neuralnet.prmdt}
\title{predict.neuralnet.prmdt}
\usage{
\method{predict}{neuralnet.prmdt}(object, newdata, type = "class", ...)
}
\description{
predict.neuralnet.prmdt
}
\references{
Stefan Fritsch, Frauke Guenther and Marvin N. Wright (2019). neuralnet: Training of Neural Networks.
R package version 1.44.2. https://CRAN.R-project.org/package=neuralnet
}
\keyword{internal}
