% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.nnet.prmdt}
\alias{predict.nnet.prmdt}
\title{predict.nnet.prmdt}
\usage{
\method{predict}{nnet.prmdt}(object, newdata, type = "class", ...)
}
\description{
predict.nnet.prmdt
}
\references{
Venables, W. N. & Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth Edition. Springer, New York. ISBN 0-387-95457-0
}
\keyword{internal}
