% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.svm.prmdt}
\alias{predict.svm.prmdt}
\title{predict.svm.prmdt}
\usage{
\method{predict}{svm.prmdt}(
  object,
  newdata,
  type = "class",
  decision.values = FALSE,
  ...,
  na.action = na.omit
)
}
\description{
predict.svm.prmdt
}
\references{
David Meyer, Evgenia Dimitriadou, Kurt Hornik, Andreas Weingessel and Friedrich Leisch (2020).
e1071: Misc Functions of the Department of Statistics, Probability Theory Group (Formerly: E1071), TU Wien. R package version 1.7-4.
https://CRAN.R-project.org/package=e1071
}
\keyword{internal}
