% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{prediction.variable.balance}
\alias{prediction.variable.balance}
\title{prediction.variable.balance}
\usage{
prediction.variable.balance(
  data,
  predict.variable,
  ylab = "Number of individuals",
  xlab = "",
  main = paste("Variable Distribution", predict.variable),
  col = NA
)
}
\arguments{
\item{data}{A data frame.}

\item{predict.variable}{Character type. The name of the variable to predict. This name must be part of the columns of the data frame.}

\item{ylab}{A character string that describes the y-axis on the graph.}

\item{xlab}{A character string that describes the x-axis on the graph.}

\item{main}{Character type. The main title of the chart.}

\item{col}{A vector that specifies the colors of the categories represented by bars within the chart.}
}
\value{
A ggplot object.
}
\description{
Function that graphs the balance of the different categories of a column of a data frame.
}
\note{
With this function we can identify if the data is balanced or not, according to the variable to be predicted.
}
\examples{

prediction.variable.balance(iris,"Species")

}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
