% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{print.indexes.prmdt}
\alias{print.indexes.prmdt}
\title{Printing prmdt index object}
\usage{
\method{print}{indexes.prmdt}(x, ...)
}
\arguments{
\item{x}{A prmdt index object}

\item{...}{optional arguments to print o format method}
}
\value{
a print of the results of a prediction model.
}
\description{
Printing prmdt index object
}
