% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{print.prediction.prmdt}
\alias{print.prediction.prmdt}
\title{Printing prmdt prediction object}
\usage{
\method{print}{prediction.prmdt}(x, ...)
}
\arguments{
\item{x}{A prmdt prediction object}

\item{...}{optional arguments to print o format method}
}
\value{
a print prediction of a model.
}
\description{
Printing prmdt prediction object
}
