% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{print.prmdt}
\alias{print.prmdt}
\title{Printing prmdt models}
\usage{
\method{print}{prmdt}(x, ...)
}
\arguments{
\item{x}{A prmdt models}

\item{...}{optional arguments to print o format method}
}
\value{
a print information of a model.
}
\description{
Printing prmdt models
}
