% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ada.prmdt}
\alias{predict.ada.prmdt}
\title{predict.ada.prmdt}
\usage{
\method{predict}{ada.prmdt}(object, newdata, type = "class",
  n.iter = NULL, ...)
}
\description{
predict.ada.prmdt
}
\keyword{internal}
