% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.bayes.prmdt}
\alias{predict.bayes.prmdt}
\title{predict.bayes.prmdt}
\usage{
\method{predict}{bayes.prmdt}(object, newdata, type = "class",
  threshold = 0.001, eps = 0, ...)
}
\description{
predict.bayes.prmdt
}
\keyword{internal}
