% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.glmnet.prmdt}
\alias{predict.glmnet.prmdt}
\title{predict.glmnet.prmdt}
\usage{
\method{predict}{glmnet.prmdt}(object, newdata, type = "class", s = NULL, ...)
}
\description{
predict.glmnet.prmdt
}
\keyword{internal}
