% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.knn.prmdt}
\alias{predict.knn.prmdt}
\title{predict.knn.prmdt}
\usage{
\method{predict}{knn.prmdt}(object, newdata, type = "class", ...)
}
\arguments{
\item{object}{a \code{\link[kknn]{train.kknn}} model object for which prediction is desired.}

\item{newdata}{an optional data frame in which to look for variables with which to predict.}

\item{type}{type of prediction 'prob' or 'class' (default).}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
a vector or matrix of predictions for knn model.
}
\description{
Return prediction for a \code{\link[kknn]{train.kknn}} model.
}
