% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.neuralnet.prmdt}
\alias{predict.neuralnet.prmdt}
\title{predict.neuralnet.prmdt}
\usage{
\method{predict}{neuralnet.prmdt}(object, newdata, type = "class", ...)
}
\description{
predict.neuralnet.prmdt
}
\keyword{internal}
