% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rpart.prmdt}
\alias{predict.rpart.prmdt}
\title{predict.rpart.prmdt}
\usage{
\method{predict}{rpart.prmdt}(object, newdata, type = "class",
  na.action = na.pass, ...)
}
\description{
predict.rpart.prmdt
}
\keyword{internal}
