% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.svm.prmdt}
\alias{predict.svm.prmdt}
\title{predict.svm.prmdt}
\usage{
\method{predict}{svm.prmdt}(
  object,
  newdata,
  type = "class",
  decision.values = FALSE,
  ...,
  na.action = na.omit
)
}
\description{
predict.svm.prmdt
}
\keyword{internal}
