% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.xgb.Booster.prmdt}
\alias{predict.xgb.Booster.prmdt}
\title{predict.xgb.Booster}
\usage{
\method{predict}{xgb.Booster.prmdt}(object, newdata, type = "class",
  missing = NA, outputmargin = FALSE, ntreelimit = NULL,
  predleaf = FALSE, predcontrib = FALSE, approxcontrib = FALSE,
  predinteraction = FALSE, reshape = FALSE, ...)
}
\description{
predict.xgb.Booster
}
\keyword{internal}
